% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmarc.R
\name{dmarc_parse_xml_report}
\alias{dmarc_parse_xml_report}
\title{Parse a DMARC XML report}
\usage{
dmarc_parse_xml_report(x)
}
\arguments{
\item{x}{path to report}
}
\description{
Parse a DMARC XML report
}
\examples{
dmarc_parse_xml_report(system.file("extdata", "report.xml", package="dmarc"))
}
