% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_dmarc_valid}
\alias{is_dmarc_valid}
\title{Test if DMARC records are valid}
\usage{
is_dmarc_valid(domains, dmarc_records)
}
\arguments{
\item{domains, dmarc_records}{character vectors of paired domains and DMARC records}
}
\description{
Test if DMARC records are valid
}
