% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-dmarc.R
\name{parse_dmarc}
\alias{parse_dmarc}
\title{Parse already retrieved DMARC records}
\usage{
parse_dmarc(domains, dmarc_records)
}
\arguments{
\item{domains, dmarc_records}{character vector of domain names and associated
DMARC records.}
}
\description{
Parse already retrieved DMARC records
}
