% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_dmarc}
\alias{read_dmarc}
\title{Read DMARC records for a set of domains}
\usage{
read_dmarc(domains)
}
\arguments{
\item{domains}{character vector of domain names}
}
\description{
Read DMARC records for a set of domains
}
